Randomize question and answer order in exams if the option is specified in the quiz exercise (#1798)
