+++ src/main/webapp/app/exercises/quiz/participate/quiz-participation.component.ts
@@ -31,6 +31,7 @@ import { ShortAnswerQuestion } from 'app/entities/quiz/short-answer-question.mod
 import { QuizQuestionType } from 'app/entities/quiz/quiz-question.model';
 import { MultipleChoiceSubmittedAnswer } from 'app/entities/quiz/multiple-choice-submitted-answer.model';
 import { DragAndDropQuestion } from 'app/entities/quiz/drag-and-drop-question.model';
+import { ArtemisQuizService } from 'app/shared/quiz/quiz.service';
 
 @Component({
     selector: 'jhi-quiz',
@@ -121,6 +122,7 @@ export class QuizParticipationComponent implements OnInit, OnDestroy {
         private quizParticipationService: QuizParticipationService,
         private translateService: TranslateService,
         private deviceService: DeviceDetectorService,
+        private quizService: ArtemisQuizService,
     ) {
         smoothscroll.polyfill();
     }
@@ -288,7 +290,7 @@ export class QuizParticipationComponent implements OnInit, OnDestroy {
         this.initQuiz();
 
         // randomize order
-        this.randomizeOrder(this.quizExercise);
+        this.quizService.randomizeOrder(this.quizExercise);
 
         // init empty submission
         this.submission = new QuizSubmission();
@@ -625,7 +627,7 @@ export class QuizParticipationComponent implements OnInit, OnDestroy {
                 this.jhiWebsocketService.enableReconnect();
 
                 // apply randomized order where necessary
-                this.randomizeOrder(this.quizExercise);
+                this.quizService.randomizeOrder(this.quizExercise);
             }
         } else {
             // quiz hasn't started yet
@@ -740,49 +742,6 @@ export class QuizParticipationComponent implements OnInit, OnDestroy {
                 // limit decimal places to 2
                 this.questionScores[submittedAnswer.quizQuestion.id] = Math.round(submittedAnswer.scoreInPoints * 100) / 100;
             }, this);
-        }
-    }
-
-    /**
-     * Randomize the order of the questions
-     * (and answerOptions or dragItems within each question)
-     * if randomizeOrder is true
-     *
-     * @param quizExercise {object} the quizExercise to randomize elements in
-     */
-    randomizeOrder(quizExercise: QuizExercise) {
-        if (quizExercise.quizQuestions) {
-            // shuffle questions
-            if (quizExercise.randomizeQuestionOrder) {
-                this.shuffle(quizExercise.quizQuestions);
-            }
-
-            // shuffle answerOptions / dragItems within questions
-            quizExercise.quizQuestions.forEach((question) => {
-                if (question.randomizeOrder) {
-                    if (question.type === QuizQuestionType.MULTIPLE_CHOICE) {
-                        this.shuffle((question as MultipleChoiceQuestion).answerOptions!);
-                    } else if (question.type === QuizQuestionType.DRAG_AND_DROP) {
-                        this.shuffle((question as DragAndDropQuestion).dragItems);
-                    } else if (question.type === QuizQuestionType.SHORT_ANSWER) {
-                    } else {
-                        console.log('Unknown question type: ' + question);
-                    }
-                }
-            }, this);
-        }
-    }
-
-    /**
-     * Shuffles array in place.
-     * @param {Array} items An array containing the items.
-     */
-    shuffle<T>(items: T[]) {
-        for (let i = items.length - 1; i > 0; i--) {
-            const pickedIndex = Math.floor(Math.random() * (i + 1));
-            const picked = items[pickedIndex];
-            items[pickedIndex] = items[i];
-            items[i] = picked;
         }
     }
